Corrected lesson_4.py. Added utilities.py
