+++ Homework 4/lesson_4.py
@@ -2,6 +2,7 @@ from database_handler import execute_query
 from flask import Flask
 from webargs.flaskparser import use_kwargs
 from validator import order_price_validator, track_validator
+from utilities import time_representation
 
 app = Flask(__name__)
 
@@ -15,21 +16,22 @@ def order_price(country) -> str:
     query = (f"SELECT BillingCountry, SUM(UnitPrice * Quantity) AS SALES FROM invoice_items LEFT JOIN invoices "
              f"ON invoice_items.InvoiceId = invoices.InvoiceId")
 
+    message = "Total sales: {}"
+
     if country:
-        country = country.title()
-        query += f" WHERE BillingCountry = '{country}'"
-        records: list = execute_query(query)
-        country_name: str = records[0][0]
-        country_sales: float = records[0][1]
-        return f"{country_name} total sales: {country_sales}"
+        query += f" WHERE BillingCountry = '{country.title()}'"
+
+    records: list = execute_query(query)
+    country_name: str = records[0][0]
+    country_sales: float = records[0][1]
+
+    if country:
+        message = message.replace("T", "{} t")
+        message = message.format(country_name, country_sales)
     else:
-        records: list = execute_query(query)
-        print(records)
-        total_sales: float = records[0][1]
-        return f"Total sales: {total_sales}"
+        message = message.format(country_sales)
 
-
-# http://127.0.0.1:5001/order-price
+    return message
 
 
 @app.route("/get-all-info-about-track")
@@ -49,13 +51,6 @@ def get_all_info_about_track(track_id) -> list:
 
     records: list = execute_query(query)
 
-    def time_representation(time_ms: int) -> str:
-        """Converts miliseconds to hours and minutes."""
-        time_h = time_ms / (1000 * 60 * 60)
-        result_time_h = int(time_h)
-        result_time_m = int((time_h - result_time_h) * 60)
-        return f"The total time is {result_time_h}h {result_time_m}m"
-
     time_query = query.replace("*", "sum(Milliseconds)")
     time: int = execute_query(time_query)[0][0]
     time: str = time_representation(time)
@@ -63,7 +58,5 @@ def get_all_info_about_track(track_id) -> list:
 
     return records
 
-
-# http://127.0.0.1:5001/get-all-info-about-track
 
 app.run(port=5001, debug=True)
